fix(settings): gate patient-only sections by role (#475 §7)
Bug #475 (§7): /settings showed all 9 sections to every role — a DOCTOR/NURSE
saw "Données médicales" (pathology/yearDiag/height), "Administratif" (NIRPP/INS),
"Moments de la journée" and "Confidentialité" (patient RGPD consents), whose
Save buttons hit /api/patient/* and would 403/404.

- Split the client form into SettingsClient.tsx (git mv from page.tsx) taking a
  `role` prop; page.tsx is now a thin Server Component reading the `x-user-role`
  header injected by the JWT middleware (matcher already covers /settings/:path*)
  — no extra round-trip. Unknown/missing role → VIEWER (fail-safe).
- SettingsClient filters SECTIONS: VIEWER (patient) sees all; NURSE/DOCTOR/ADMIN
  no longer see medicalData/administrative/dayMoments/privacy.
- tests/pages/settings.test.tsx: render SettingsClient with a role (VIEWER keeps
  the existing 8-section assertions) + 2 new tests (DOCTOR hides patient-only,
  VIEWER shows them).

tsc + eslint clean; settings page tests 10/10 green.

Closes #475.
